Extract shared test utilities and add draft-15 test constants
Refactor only — no production behavior change.

- Extract ~530 lines of reusable test fixtures (TestSession, mock helpers,
  ReceiveWebTransportSession, etc.) from quic_spdy_session_test.cc into
  quic_spdy_session_test_utils.h for sharing across draft-15 test files
- Add QuicSpdySessionPeer::EnableWebTransport(session, versions) overload
  to allow tests to specify exact version sets
- Add draft15_constants.h with spec codepoints for test assertions
  (SETTINGS values, error codes, capsule types from Sections 5, 9.5)
